requirements.txt
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-admin-bootstrapped==2.4.0
django-bootstrap3==5.4.0
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.4.0
django-datetime-widget==0.9.3
django-endless-pagination==2.0
django-form-utils==1.0.2
django-pagination==1.0.7
django-simple-captcha==0.4.5
django-tastypie==0.12.1
django-toolbelt==0.0.1
djangorestframework==3.1.1
gunicorn==19.3.0
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
psycopg2==2.6
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.2
six==1.9.0
static3==0.5.1
